+++ sunhome-thread/src/main/java/com/sunhome/framework/thread/base/chapter1/package-info.java
@@ -1,0 +1,5 @@
+package com.sunhome.framework.thread.base.chapter1;
+/**
+ * 多线程基础篇-了解线程的原理
+ *
+ */
