+++ sunhome-thread/src/main/java/com/sunhome/framework/thread/base/chapter2/package-info.java
@@ -1,0 +1,9 @@
+package com.sunhome.framework.thread.base.chapter2;
+
+/**
+ * 多线程基础篇-走进锁机制
+ *
+ * 层面停留在使用层面
+ * 后期补充jvm原理
+ *
+ */
