--- sunhome-thread/src/main/java/com/sunhome/framework/thread/core/chapter1/package-info.java
@@ -1,5 +1,0 @@
-package com.sunhome.framework.thread.core.chapter1;
-/**
- * 多线程基础篇-了解线程的原理
- *
- */
