--- sunhome-thread/src/main/java/com/sunhome/framework/thread/core/chapter2/package-info.java
@@ -1,9 +1,0 @@
-package com.sunhome.framework.thread.core.chapter2;
-
-/**
- * 多线程基础篇-走进锁机制
- *
- * 层面停留在使用层面
- * 后期补充jvm原理
- *
- */
